Divide módulo target skill em buff e debuff
